requirements.txt
streamlit
numpy
pillow
vision-agent==0.2.233
opencv-python
pillow-heif
torch==2.0.1
diffusers==0.21.4
transformers==4.31.0
accelerate==0.21.0
